backend/requirements.txt
django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
langchain==0.1.0
langchain-openai==0.0.5
pinecone-client==3.0.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
python-multipart==0.0.6
pypdf==4.0.1